Added scope named "unspecified" and changed empty scopes array to use this scope.
In the OpenAPI v3 specification, Security Scheme Object requires a map of scopes, but doesn't specify "non-empty." The Security Requirement Object strongly implies that a non-empty array of scopes is required, but again doesn't specify. KaiZen OpenAPI Editor treats the empty scopes array as invalid.

I'll take this up with the OpenAPI specification working group. If empty scopes is allowed in security scheme, it must be allowed in requirements. Conversely, if it's optional in security requirement object, it must be optional in the security scheme.\

In the meantime, I've added this "unspecified" scope and referenced that from security requirements. This gets around the issue for now.